pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
name = "mlb-data-platform"
dynamic = ["version"]
description = "Kubernetes-native MLB Stats API data platform with PySpark ETL pipelines"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Apache-2.0" }
authors = [
    { name = "Nikolaus Schuetz", email = "nikolauspschuetz@gmail.com" }
]
keywords = [
    "mlb",
    "baseball",
    "etl",
    "data-platform",
    "pyspark",
    "kubernetes",
    "argo-workflows",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "pymlb-statsapi>=1.0.0",
    "pyspark>=3.5.0",
    "delta-spark>=3.0.0",
    "pydeequ>=1.2.0",  # AWS data quality library for PySpark
    "psycopg[binary]>=3.1.0",
    "psycopg-pool>=3.1.0",
    "sqlmodel>=0.0.14",  # SQLAlchemy 2.0 + Pydantic for ORM
    "redis>=5.0.0",
    "minio>=7.2.0",
    "pyyaml>=6.0.1",
    "typer>=0.9.0",
    "rich>=13.7.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "fastavro>=1.9.0",  # Use fastavro instead of avro-python3
    "httpx>=0.26.0",
    "tenacity>=8.2.3",
    "python-json-logger>=2.0.7",
    "beautifulsoup4>=4.12.0",  # HTML parsing for Seamheads scraper
]

[dependency-groups]
dev = [
    "pytest>=7.4.3",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.1",
    "behave>=1.2.6",
    "ruff>=0.1.9",
    "mypy>=1.8.0",
    "bandit>=1.7.6",
    "pre-commit>=3.6.0",
    "ipython>=8.19.0",
    "ipykernel>=6.28.0",
]

docs = [
    "sphinx>=7.2.6",
    "sphinx-rtd-theme>=2.0.0",
    "myst-parser>=2.0.0",
]

[project.scripts]
mlb-etl = "mlb_data_platform.cli:app"

[project.urls]
Homepage = "https://github.com/power-edge/mlb_statsapi_data_platform"
Repository = "https://github.com/power-edge/mlb_statsapi_data_platform"
Issues = "https://github.com/power-edge/mlb_statsapi_data_platform/issues"

[tool.hatch.version]
source = "vcs"

[tool.hatch.build.hooks.vcs]
version-file = "src/mlb_data_platform/_version.py"

[tool.hatch.build.targets.wheel]
packages = ["src/mlb_data_platform"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
    "ARG",  # flake8-unused-arguments
    "SIM",  # flake8-simplify
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
    "ARG001",  # unused function argument
    "ARG002",  # unused method argument
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # Allow unused imports in __init__.py
"tests/**/*.py" = ["ARG", "S101"]  # Allow assert in tests

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "--cov=mlb_data_platform",
    "--cov-report=term-missing",
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true

[[tool.mypy.overrides]]
module = [
    "pyspark.*",
    "pydeequ.*",
    "minio.*",
    "avro.*",
    "fastavro.*",
]
ignore_missing_imports = true

[tool.bandit]
exclude_dirs = ["tests", "test_*"]
skips = ["B101", "B601"]

[tool.coverage.run]
source = ["mlb_data_platform"]
omit = ["tests/*", "*/__pycache__/*", "*/site-packages/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
